# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/customers/me": {
      "get": {
        "operationId": "CustomersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user details"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/customers/me/password": {
      "patch": {
        "operationId": "CustomersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/api/news": {
      "get": {
        "operationId": "NewsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of news"
          }
        },
        "summary": "List news (paginated)",
        "tags": [
          "News"
        ]
      },
      "post": {
        "operationId": "NewsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNewsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "News created"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create news (emits news.created event)",
        "tags": [
          "News"
        ]
      }
    },
...
    "/api/news/slug/{slug}": {
      "get": {
        "operationId": "NewsController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News details"
          }
        },
        "summary": "Get news by slug",
        "tags": [
          "News"
        ]
      }
    },
    "/api/news/{id}": {
      "get": {
        "operationId": "NewsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News details"
          }
        },
        "summary": "Get news by ID",
        "tags": [
          "News"
        ]
      },
      "put": {
        "operationId": "NewsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNewsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "News updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update news (emits news.updated event)",
        "tags": [
          "News"
        ]
      },
      "delete": {
        "operationId": "NewsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete news (emits news.deleted event)",
        "tags": [
          "News"
        ]
      }
    },
    "/api/news/{id}/image": {
      "post": {
        "operationId": "NewsController_uploadImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image uploaded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload news image",
        "tags": [
          "News"
        ]
      }
    },
...
    "/api/contacts": {
      "post": {
        "operationId": "ContactsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created"
          }
        },
        "summary": "Create contact (emits contact.created event)",
        "tags": [
          "Contacts"
        ]
      },
      "get": {
        "operationId": "ContactsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of contacts"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List contacts (paginated)",
        "tags": [
          "Contacts"
        ]
      }
    },
    "/api/contacts/stats": {
      "get": {
        "operationId": "ContactsController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Contact stats"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get contact statistics",
        "tags": [
          "Contacts"
        ]
      }
    },
    "/api/contacts/{id}": {
      "get": {
        "operationId": "ContactsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact details"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get contact by ID",
        "tags": [
          "Contacts"
        ]
      },
      "patch": {
        "operationId": "ContactsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update contact (emits contact.updated event)",
        "tags": [
          "Contacts"
        ]
      },
      "delete": {
        "operationId": "ContactsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete contact",
        "tags": [
          "Contacts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Baccouche Motors API Gateway",
    "description": "API Gateway for Baccouche Motors microservices.\n\n## Event-Driven Design (RabbitMQ)\nThis system follows a strict **Event-Driven Design** pattern:\n- **Domain events** are emitted on create/update/delete operations\n- **Consumers** react asynchronously (emails, stats, audit)\n\n| Event | Producer | Consumers |\n|-------|----------|-----------|\n| `auth.forgot-password-requested` | Auth | Notifications (email) |\n| `car.created` / `updated` / `deleted` | Cars | Admin (stats) |\n| `news.created` / `updated` / `deleted` | News | Admin (stats) |\n| `contact.created` / `updated` | Contacts | Notifications, Admin |\n| `test-drive.created` / `status-changed` | Test Drives | Notifications, Admin |\n\n**Flow**: Gateway → RabbitMQ RPC (sync) → Microservice → emits to `baccouche.events` exchange → Consumers (Notifications, Admin)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Current password"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
      "CreateNewsDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "example": "2025-01-15"
          },
          "image": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          }
        },
        "required": [
          "slug",
          "title",
          "excerpt",
          "content",
          "date"
        ]
      },
      "UpdateNewsDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          }
        }
      },
...
      "CreateContactDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "phone": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "subject",
          "message"
        ]
      },
      "UpdateContactDto": {
        "type": "object",
        "properties": {
          "read": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}